+++ arc.h
@@ -10,13 +10,13 @@ class Arc {
  }
 
  Arc(uint32_t src_id, uint32_t dst_id, int32_t capacity, int32_t flw,
-     int32_t cst, Arc* rvrd_arc): src_node_id(src_id), dst_node_id(dst_id),
+     int64_t cst, Arc* rvrd_arc): src_node_id(src_id), dst_node_id(dst_id),
     cap(capacity), flow(flw), cost(cst), reverse_arc(rvrd_arc) {
  }
 
   int32_t get_cap();
   int32_t get_flow();
-  int32_t get_cost();
+  int64_t get_cost();
   uint32_t get_src_node_id();
   uint32_t get_dst_node_id();
   Arc* get_reverse_arc();
@@ -26,7 +26,7 @@ class Arc {
   uint32_t dst_node_id;
   int32_t cap;
   int32_t flow;
-  int32_t cost;
+  int64_t cost;
   Arc* reverse_arc;
 
 };
